Tooltip implementation changed
